refactor(types): move workflow types to @jaicome/zatca-core
Client code should not depend on @jaicome/zatca-server. Move OnboardResult,
SingleReportResult, BatchReportResult, and ClientInvoiceRecord to
@jaicome/zatca-core. Server re-exports them for backward compatibility.
