+++ packages/zatca-server/src/api/types.ts
@@ -29,30 +29,5 @@ export interface InvoiceResponse {
   clearanceStatus?: string;
 }
 
-// --- Workflow Result Types ---
-
-export type OnboardResult = {
-  previousInvoiceHash: string;
-  invoiceCounterStart: number;
-};
-
-export type SingleReportResult = {
-  invoiceSerialNumber: string;
-  success: boolean;
-  invoiceHash: string;
-  reportingStatus?: string;
-  clearanceStatus?: string;
-  error?: string;
-};
-
-export type BatchReportResult = SingleReportResult[];
-
-export interface ClientInvoiceRecord {
-  invoiceSerialNumber: string;
-  invoiceCounterNumber: number;
-  previousInvoiceHash: string;
-  props: import("@jaicome/zatca-core").ZATCAInvoiceProps;
-  status: "pending" | "reported" | "failed";
-  serverResponse?: SingleReportResult;
-  createdAt: string;
-}
+ // Re-exported from @jaicome/zatca-core for backward compatibility
+export type { OnboardResult, SingleReportResult, BatchReportResult, ClientInvoiceRecord } from "@jaicome/zatca-core";
